Fixed a bunch of errors

+++ server/src/main/java/service/LogoutService.java
@@ -2,6 +2,7 @@ package service;
 
 import dataaccess.AuthDAO;
 import dataaccess.DataAccessException;
+import exception.DatabaseErrorException;
 import exception.NotAuthenticatedException;
 import exception.ResponseException;
 
@@ -13,10 +14,16 @@ public class LogoutService {
     }
 
     public void logout(String authToken) throws ResponseException {
+        if (authToken == null) {
+            throw new NotAuthenticatedException();
+        }
         try {
+            if (authDAO.getAuthByToken(authToken) == null) {
+                throw new NotAuthenticatedException();
+            }
             authDAO.deleteAuthByToken(authToken);
         } catch (DataAccessException e) {
-            throw new NotAuthenticatedException("Error: invalid auth token");
+            throw new DatabaseErrorException("could not delete auth");
         }
     }
